+++ src/views/snippets/configuration-skin-default.js
@@ -7,7 +7,7 @@ module.exports = {
         borderRadius: '.25rem',
         borderWidth: '1px',
         boxShadow: '0 4px 6px -1px rgb(0 0 0 / 0.1), 0 2px 4px -2px rgb(0 0 0 / 0.1);',
-        themes: {
+        themes: { // Valid themes: light, dark
             light: {
                 bgDefault: colors.gray["100"],
                 bgContent: colors.white,